Allow MultiConfigParser get from mutliple sections
MultiConfigParser.get() takes a section name and a list of option names.

In order to support aliases of options which span multiple groups, we
need to be able to look up from multiple sections. We can easily do this
by changing MultiConfigParser.get() to take a list of (section, name)
tuples.

Change-Id: I50656c021669ccb947ceb56dfcf37d3fffcba194