Change the name of the project to oslo.config
Again discussed here:

  http://lists.openstack.org/pipermail/openstack-dev/2013-March/thread.html#6283
  http://lists.openstack.org/pipermail/openstack-dev/2013-March/thread.html#6138

The rationale is that the use of namespace packages is likely to
increase and there is apparently some consensus that the convention
should be to use period to separate the namespace package name from
the subpackage name - i.e. call it oslo.config. The zope.interface is
an example of this convention, but PasteDepoy is a counter-example.

Oslo is going to have a family of libraries, so it'd be much nicer if
they followed what the future Python convention is likely to be.

This is a disruptive change for packagers and the benefit isn't huge,
but pairing it with the also disruptive change to the versioning
scheme makes it somewhat more acceptable.

Change-Id: I63e96a8b48b166b07f1c8e79b6eaffb6234cc751