Re-work test_cli to use testscenarios.
The cli tests were already organized in a manner that lets testscenarios
do what it does. Reformat the code a little bit to be a single test plus
a list of declarative combinations of scenarios and let testscenarios
do the work of turning that into a set of test cases.

Change-Id: I4e10d9eaeb313c724f0012babac465e37be4fa3d