Add auto-create support for OptGroup instances
Currently, it is possible to auto-create groups by passing a group name
to register_opt or register_opts. However, it is not possible to
auto-register groups when a OptGroup instance is passed instead.

This patch adds support for that to the _get_group method. There are
some advantages introduced by this patch:

    * It keeps consistency of oslo.config's public API.
    * It allow users to auto-register a group a group and
    re-use the instance:
        group = cfg.OptGroup("test")
        cfg.register_opts(..., group=group)
        # do something with the group
    * It allows users to auto-register groups with descriptions.

DocImpact

Change-Id: I9b9a74688c906d6bdea9c42044b47dc52e48c19d