Update requirements
This is the fabulous result of running the update.py script from
openstack/requirements with this repository as the target. It therefore
updates all dependencies, with nice version numbers that this repository
have to use to work. It also fixes case on some dependencies.
This update also brings a brand shiny new setup.py automatically handled
and generate by this same openstack/requirements repository.

Change-Id: I22ec396e5fa802c1617514aa6df3ed82f05271b0