Remove explicit version from setup.cfg
With the explicit "pre version" in setup.cfg, PBR generates our
version as 1.2.0.a1.g2748574 and, without it, it generates
1.2.0a4.1.g2748574.

which in turn causes nova to complain about:

 VersionConflict: (oslo.config 1.2.0.a1.g2748574
 (/opt/stack/new/oslo.config),
 Requirement.parse('oslo.config>=1.2.0a3'))

when trying to load extensions with stevedore.

I'm not sure why this is suddenly an issue after 1.2.0a4 has been
tagged, but removing this version is the right thing to do anyway
since we want to rely on tags for all versioning info.

Change-Id: I9090b8401a43befd73a82f9bb78e0b6165619c7c