Raises error if duplicate keys found in DictOpt
DictOpt didn't report when the input values had duplicate keys.
This would be a problem if the user had put in duplicate keys in
the conf file and the service didn't work as expected.
(c.f) Neutrons provider network mapping options
This patch raises a ConfigFileValueError when duplicate keys are found.

Fixes bug #1228995

Change-Id: I91e1bd0eda4ed4fd6c7eb6ea5d4a89d583d1305e