Refactors _Namespace to make the code clearer
The original implentation of _Namespace._get_cli_value relied on the
fact that _Namespace.default doesn't exist and accessing it would raise
an AttributeError. This attribute was being caught for another reason so
the logic just worked.

The more correct thing to do would have been to just raise a KeyError.
This change does that and removes the need to catch an AttributeErrors.

Change-Id: I963b31ea6ea8ab5180314be7cd2450395263e1af
Closes-bug: #1284968