Add a list_all_sections method to ConfigOpts
Currently there is no way of getting the list of all sections defined
in the files, without accessing private parts of the library. This patch
adds a method that makes it possible.

Why do we need the list of all sections? Because we have separate
sections for the panels and dashboards, and it's impossible to tell
what panels and dashboards are available before reading the
configuration. Forcing the users to list all the sections in a separate
place is information duplication, error-prone and bad usability. So
the way we do it is listing all the sections, and then registering the
groups for the ones that we care about.

Change-Id: Id7b8f8bd2423123b4c3745d78fa857a5d94e02f0