Add exception handling for entry points
oslo-config-generator is used to generate config sample iterating all
entry points provided to it.
But if some error appears, return code will be absent.
It is critical to have it running in CI.
So, add new func that will be used for error handling and provide it
to stevedore that suppresses all errors by default.

Change-Id: Ia78cc0d4e8ab842cd0e279713badd68606a67a58
Closes-Bug: #1419458