Add ability to deprecate opts for removal
Previously opts could only be deprecated by renaming them, but there
are circumstances where we may want to remove an opt completely, and
we need a way to notify users before this happens.  This change adds
a deprecated_for_removal parameter to the Opt constructor that can
be used to mark opts for which this is planned.  When such opts are
referenced it will log a message, but only if the default value has
been overridden.

Change-Id: I0165895d58530f5fce58cbf8a1bbe2acd90e9063