Expose min and max to IntOpt
The IntOpt class utilizes the types.Integer class which has
parameters for setting the minimum and maximum value.  These min
and max values are not exposed through IntOpt and should be ideally.

In the docstring help for cfg.py, it gives an example of creating a
PortType instance of types.Integer to create a new type with a range.
Rather than projects having to define this new instance, seems better
to just expose the min and max to IntOpt.

Another advantage of adding min and max to IntOpt is the ability to
generate a useful sample config file that displays the range of
valid integer values.

Change-Id: Icce7b6799061711ea512d60facc57bf7d6f6c9cc