Make Integer type class honor zero for min/max
The Integer type class has an option for setting a min or max value for the
integer. But the class would not recognize zero as a valid value and not
check it.

Make the Integer class honor zero as a valid value for min and/or max.

Add test cases for using zero for min and/or max.

Change-Id: I395dea133d2a92e5ca4fab913aa17483a38f7ba1
Closes-Bug: #1489688