Case insensitiveness when comparing choices values
This path is adding a flag called "ignore_case" for the StrOpt type which
allow case insensitiveness for 'choices' and 'regex' values. The flag
receives a boolean value and if True, the case differences (uppercase
vs. lowercase) between the values of choices will be ignored. It defaults
to False.

This flag is important because it helps with backward compability of a
StrOpt configuration that before accepted case insensitive and now is
using 'choices'

Change-Id: I5f68eee59588dd2be62e11491ad73d452bc9fc07