Allow method set_override with parameter override=None for all types
Method set_override will check parameter override's type when
enforce_type=True, If parameter override with value None, it doesn't
work for some types like IntOpt, or doesn't work properly for StrOpt.
For IntOpt, it raise TypeError, and For StrOpt, it convert None to
string 'None', that's not we expect.

Closes-Bug: #1515230
Change-Id: I42ceeccf59111057b5a92c849c9bb5f83255b324