Handle None section name in MultiOptConfigParser#get
A lot of oslo.config accepts None to mean the 'DEFAULT' section, but
MultiOptConfigParser#get doesn't. Establish this as an invariant on
MultiOptConfigParser.

Note _check_deprecated already did this but I've removed the logic here
as it's only called by _get (and tests) which now handles the job for
it. This means one of the tests was redundant so I've removed it.
Altering _check_deprecated should be OK because it's a private method so
we don't guarantee API stability.

Change-Id: I1a983df12ca550cd22d8edb806c3acc71313a959