Add 'mutable' param to Opt
re: https://etherpad.openstack.org/p/mitaka-cross-project-dynamic-config-services

oslo.config already offers a reload_config_files method. This patch
provides a new mutate_config_files method which is similar but respects
a new 'mutable' flag to Opt, allowing services to mark which opts may be
reloaded. If a new value for an immutable opt is detected, a warning is
issued.

This method also returns details of which keys have changed, which
allows users to only refresh affected resources. Refreshing some
resources, EG RabbitMQ or MySQL connections, unnecessarily can have an
end-user impact.

This is the second attempt at this feature, the previous attempt had to
be reverted because it included a callback hook which it turns out we
don't want.

Change-Id: Ie9e9a895da1a3fc62580c1febb7cd9f965bbc0f5