Add support for a deprecation reason
It is possible to deprecate an option, but there's no clear way to
provide a reason for this deprecation. This leaves the user with a
decision to make. Nova, for example, adopts the approach of adding
the reason to the end of the help message like so:

    DEPRECATED: use standard logging configuration.

This is a decision that the user shouldn't have to make. Add a
'deprecation_reason' parameter to allow users to specify a reason
while ensuring output will be displayed uniformly.

Change-Id: I71330f4d0eae1f48ed2ac842569d692081495637