generator: ignore doubles
There could be cases that a consuming project lists its config
options multiple times for the sample config file generation.
The result is that these config options get emitted multiple times
in the sample file. The generator didn't recognize this.

This commit removes the doubles from the read listings. The generator
will emit each config option (in a group in a namespace) only once.

Change-Id: Ifadc0b260ed6ab686131ac06ef74cf1f5b3dfe71