Warn if immutable options change
Log a warning if, after config files have been reloaded with
_mutate_config_files, immutable option values do not match between the
original namespace and the mutated one. Immutable options will always
be returned from the original namespace but presumably an ops person
changed the option value in the config file expecting it to be
reloaded, so we should warn them we're unable to reload this option.

Change-Id: I0212904a0a51f1b10a17c4cb46598d7bd55c947e