Report mutated options
Suppose we make the MySQL password mutable. This could be quite useful
so that the password can be rotated without having to restart
OpenStack services. Mutating this involves opening a new MySQL
connection, switching all new requests to use the new connection and
closing the old one. Doing all this work just in case the password was
changed is wasteful. Therefore a mechanism to easily know what has
changed each time the config is mutated is required.

This patch briefly caches the old _mutate_ns so that it can be diffed
against the new one. _mutate_config_files will now return this
structure. As it had no return value before, this should present no
back compatibility issues, even if this chain merges across two
versions of oslo.config.

Change-Id: I861a98da7025ce44da8d424638036d00aa9eaa4b