treat option group names as section headers when documenting them
The previous iteration of the config option doc generation code inserted
"object description" nodes, typically used for documenting things like
functions and classes. Those nodes don't look great and don't appear in
the table of contents. Replace the directive implementation with one
that uses a subheading to ensure that the name appears in the table of
contents, while also recording it for use later with the option
directive.

Change-Id: I8baafdeef1a86b66a82ede8de4a3d9515c69a4a6
Signed-off-by: Doug Hellmann <doug@doughellmann.com>