Add config_dirs property with a list of directories
config_dir currenly contains just the last directory. We need to
add a new property config_dirs which can contain a list of all
the directories on the command line. This will help fix a problem
with find_file as well since currently it was only looking at the
last directory and not all directories.

Closes-Bug: 1544334
Change-Id: Ibd0566f11df62da031afb128c9687c5e8c7b27ae