Avoid duplicate sections in generator output
Because option groups can be specified as either bare strings or
OptGroup objects, it is possible to end up with duplicate sections
in the resulting output file.  This is confusing, and can cause
problems for config management systems so we should avoid doing it.

This change avoids the dupes by always keying the group list on
the group name, but including the OptGroup object in the data
when available so the output code can make use of the extra
information in an OptGroup.

Co-Authored-By: Alexis Lee <lxsli@hpe.com>
Change-Id: I2cd8ae87a1f573b5837d292abcf7babf5d88bee5
Closes-Bug: 1568820