Add support for a deprecated version
It is possible to deprecate an option, but there's no clear way to
state when this deprecation occured and when the option will actually
go away. This leaves the user with a decision to make. Nova, for
example, frequently includes the release that the option was removed in
as part of the help message like so:

    "memcached_servers" opt is deprecated in Mitaka. In Newton release
    oslo.cache config options should be used as this option will be
    removed.

This is a decision that the user shouldn't have to make. Add a
'deprecated_since' parameter to allow users to specify this information
in a uniform manner. The parameter allows for any string value.

A future work item would add a hacking or tox check to compare this
parameter against the expected deprecation version. For example, to
ensure all Nova parameters deprecated in '13.0' should be removed in
'14.0'. This will be done on a per-project basis.

Change-Id: Ic894358006606f123e31611f068d9b6192d42616