Add 'DEPRECATED: ' prefix for deprecated opts
'f5e2fab' added support for a deprecated reason. However, there is
are a number of cases where users still prefix option help strings
with 'DEPRECATED: ', presumably as this provides a more visible
indicator of the deprecation.

Rather than leaving it up to users to do this manually, prefix
these strings automatically. A check is included to ensure duplication
does not occur.

Change-Id: I79c20d31f664d01534b6798b134b2c3fd363faec