generator: format string default value for List type properly
List type allows to pass string values as default values, which are then
handled by splitting on commas. For formatting matters, if the default
value is of string type, there is no special conversion needed, and we
can reuse the value as-is.

Change-Id: Ie9ecdfc064e886f6138b1fa9e839d8f95a7df339
Closes-Bug: #1586366
(cherry picked from commit 75e1c306630378673861d52a34de51e869183062)