Handle getattrs to ConfigOpts for deprecated opts
There are libraries and tools which offer their config interface as
a stable backwards compatible interface both in code and as user
facing interface. A vital part of being able to make changes to this
is to handle deprecation and moving options. While oslo.config offers
a fairly rich deprecation mechanism it only exposes itself on the
user facing side. If a project wishes to deprecate an option for
code consumers who wish to use the config definitions externally they
will break by any change that deprecates an option. This commit fixes
this issue by doing a lookup on the deprecated options as a fallback
on a getattr. This way if an option is not found all the deprecated
options are checked, if a matching deprecated opt is found a warning
is logged and it's value is used.

Change-Id: I6e28c25f04273f7def486fadd541babc8cf423cb