Add 'summarize' option
Projects like Nova have been putting a lot of work into centralizing
and improving documentation of configuration options. One unfortunate
side effect of this change is the huge increase in the size of the
'nova.conf' configuration file, where, for every one line configuration
option there will be many lines of documentation.

Seeing as this documentation is already published using Sphinx, there
may be little reason to pollute the configuration file with this level
of detail. Add a configuration option that, when enabled, will ensure
only the the first paragraph (a summary) of each opt or group will be
output to the generated config file. The generated Sphinx documents
remain unchanged.

Change-Id: I5fed21a333b0ac75eef4ce17e5b377f956abde7b