Fixup list types handling tuples
The addition to enforce_type now tries to convert the
value (even if enforcing is off) which can result in
types the list class was not handling now being handled.

One example is someone setting an override for a tuple
go through the same code path as for lists.

Related-Bug: #1635717

Co-Authored-By: Adam Harwell <flux.adam@gmail.com>
Change-Id: Icdce83d731642724ea3ce6920b84c25d61bd63c5