Handle both - and _ forms of deprecated opts
While config file opts cannot have a - in their name, CLI opts can
and the current deprecation code doesn't handle that case properly.
Additionally, because oslo.config has some magic to translate one
to the other it is possible we have existing users relying on the
wrong behavior.  For this reason, let's add both forms of the opt
when adding deprecated opts.  This will allow both to work as
expected.

The config generator is updated to ignore the - form of deprecated
opts since they won't work in a config file anyway.

I'm unsure how to filter the extra opts out of the help text, but
since both forms will work in that case it's a minor cosmetic
issue.

Change-Id: I4d2584b33d97fe00a427b992682d9eac391725a2
Closes-Bug: 1279973