Use Sphinx 1.5 warning-is-error
With pbr 2.0 and Sphinx 1.5, the setting for treat sphinx warnings as
errors is setting warning-is-error in build_sphinx section. Migrate
the setting from the old warnerrors one.

There are a number of issues that need to resolved to make this work.
Most of these trivial enough but unfortunately the history document
contains issue due to its inclusion of ChangeLog, which is based on
commit message summary lines and contains some invalid rST.  There
doesn't appear to be anyway to edit there retroactively nor disable
warnings on a specific files, thus the removal.

Change-Id: Ied79fc95d3711f8f3880e3e07e20f2c69b9a36ab