Don't set enforce_type if we are using the default
The test fixture for oslo.config always sets enforce_type parameter in
set_default, however because of debtcollector, this means it also
always emits a deprecation warning as well, even when the caller never
set this parameter.

oslo.config's fixtures should not trigger deprecation warnings during
normal usage. So be extra careful and fix that.

Change-Id: I5471b5d164ff40785e07b5f66e1f4673cf716971