fix formatting of literal default values with spaces
When the RST parser encounters a literal value with spaces before the
closing backticks, it misinterprets that as an unclosed literal
expression. Fix the logic that used to add quotes around such lines,
both to make the values more visible to readers of the documentation
and to fix the parsing issue.

Change-Id: I1bfd675d6611c59f4b3cc4816d48dbb199340bec
Signed-off-by: Doug Hellmann <doug@doughellmann.com>