Make help message include choices information for StrOpt
Config option `StrOpt` accept choices parameter to indicate
valid values and include comments "# Allowed values:" when generating
sample config file, so it's unnecessary to add allowed values in hep
string of the config option. But it doesn't include choices information
when printing help for the config option. This commit makes consistency
for both of sample config file and help output of command line.

Closes-Bug:1727683

Change-Id: I962e49c81bdf44043a6e1233408a6ee5b883acde