sphinxext: Start parsing 'Opt.help' as rST
Users expect this to be parsed as rST and write their docstrings
accordingly. This has the potential to introduce warnings for users with
improperly formatted rST and these warnings can be promoted to errors if
'sphinx-build' is used with the '-W' option. As a result, we disable the
'warning-is-error' logger for these options. We may wish to change this
behavior in the future.

It is not really possible to test this yet as the output wouldn't look
much different. In addition, the error messages generated are rather
unhelpful. Both of these can be changed in a future modification.

Change-Id: Ic6c2dcbe7823dd3fdc71db8dc5afab2d604559e3
Closes-Bug: #1755783
(cherry picked from commit 03ff144f0bf386982d379d18aeb9183146a05fd5)