Add bindep.txt file to prevent fallback to generic list
Jobs will look for bindep.txt in each repo to get a list of required
system packages to be installed. If that file is not found, it will
fallback to a generic bindep.txt file that contains many things that
are not needed for running oslo.utils jobs (things such as mysql and
postgres packages).

This adds a basic bindep.txt file with just a few packages to speed
up job setup times and prevent unnecessary work in the gate.

Change-Id: I9be96f4d5fc5833e29f89e96df0099cbc9f3b5fa