Remove MultiConfigParser
This has been deprecated for quite a while and some of the upcoming
work to be done in oslo.config will be easier if we don't have to
deal with this unused parser too.

The class is still currently in use in networking-cisco, but in
change https://review.openstack.org/554617 we have provided a fix
that removes their reliance on it.  They are planning to remove it
entirely in Rocky and are only testing against a capped version in
Queens so we are not going to wait on them.

Change-Id: Id6ae40311f967e172e45bcb0e61812365a72618b