Simple config validator
This is a very basic config validator that just reads the provided
config file and compares it with the machine readable sample config
data from the config generator to find any options that are not
defined in the project. It will also warn about deprecated options
in use in the provided file.

Change-Id: I410b16ffcfb069140bd51dffaba7d56eada1a7e9