generator: Pass conf to _OptFormatter
In preparation for adding another conf option that would have needed to
be kwarg'd all the way down the call stack to the .format method, this
patch removes the existing `minimal` and `summarize` kwargs from the
methods in that call stack and instead makes them instance attributes of
generator._OptFormatter, which is now initialized with the whole
ConfigOpts object rather than those individual options.

This is just a refactor. There is no functional change (assuming nobody
is being naughty and consuming the private _OptFormatter class).

Change-Id: Ide0f1c6d79d506b979b54e6446786ea9ff51c8d5